+++ springMentoring/src/main/java/com/foxminded/vitaliifedan/task7/dao/GroupDao.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
-public interface GroupDao extends CrudDao<Group, Integer> {
+public interface GroupDao extends CrudDao<Group, Long> {
     Optional<Group> findByGroupName(Connection connection, String name) throws SQLException;
 
     List<Group> findAllGroupsWithLessOrEqualStudentCount(Connection connection, int count) throws SQLException;
